Change pattern of shading relocation
Related to issue #16:
maven-shade-plugin changes values of string literals in shaded
libraries, so every literal string beginning with a pattern would be
relocated.
In this case in clickhouse-jdbc:
The string "comment" was relocated to "com.shaded.liquibase.comment"

The bug https://issues.apache.org/jira/browse/MSHADE-156 was filed on
the plugin but does not seem on its way to be fixed, so we try the
following workaround: Apply a stronger pattern.